requirements.txt
Flask==1.1.2
Pillow==7.2.0
requests==2.24.0
numpy==1.18.5
json5==0.9.5